requirements.txt
annotated-types==0.6.0
asgiref==3.7.2
Django==4.2.10
django-ninja==1.1.0
psycopg2-binary==2.9.9
pydantic==2.6.1
pydantic_core==2.16.2
sqlparse==0.4.4
typing_extensions==4.9.0
